Wind sensor fixes

+++ src/sensors/scripts/winddirection.py
@@ -16,7 +16,7 @@ class WindDirectionNode:
     def __init__(self):
         self.relativeWindMarkerPub = rospy.Publisher("/relative_wind_marker", Marker, queue_size=10)
         self.windDirectionRawSub = rospy.Subscriber("/relative_wind_direction/raw", Int32, self.updateRelativeWind, queue_size=10)
-        self.windDirectionPub = rospy.Publisher("/relative_wind_direction", Int32, queue_size=10)
+        self.windDirectionPub = rospy.Publisher("/relative_wind_direction", Float32, queue_size=10)
         self.tfBroadcaster = tf.TransformBroadcaster()
 
         self.relativeWindMarker = Marker()
@@ -51,7 +51,7 @@ class WindDirectionNode:
 
         windAngle = atan2(self.sinval, self.cosval)
 
-        self.windDirectionPub.publish(Int32(degrees(windAngle) + self.offset))
+        self.windDirectionPub.publish(Float32(degrees(windAngle) + self.offset))
 
         relativeWindQat = tf.transformations.quaternion_from_euler(0, 0, (windAngle + radians(self.offset)))
         self.relativeWindMarker.pose.orientation.x = relativeWindQat[0]
